added clients, contact messages, donation requests and app settings modules.

+++ resources/views/contact-messages/show.blade.php
@@ -3,7 +3,7 @@
 
 
 @section('page_title')
-Categories
+Contact Message Details
 @endsection
 
 
@@ -15,70 +15,62 @@ Categories
       <!-- Default box -->                 
          <div class="card">
             <div class="card-header">
-              <h3 class="card-title">List of all categories</h3>
+              <h3 class="card-title">Details of message</h3>
             </div>
             <!-- /.card-header -->
             <div class="card-body">
 
 
               @include('flash::message')
-
-              
-              <a href="{{url(route('category.create'))}}">
-                <button type="button" class="btn btn-success" style="float: right; font-weight:bold;">Add   <i class="fas fa-plus"></i></button>
-              </a>
-
               <br><br>
 
-            @if(count($records))
-             <table id="example1" class="table table-bordered table-striped table-hover">
+            @if($record)
+             <table class="table table-bordered table-striped table-hover">
                 <thead>
                 <tr>
-                  <th>#</th>
                   <th>ID</th>
                   <th>Name</th>
+                  <th>Phone</th>
+                  <th>Subject</th>
+                  <th>Body</th>
                   <th>Created_at</th>
                   <th>Updated_at</th>
-                  <th>Edit</th>
                   <th>Delete</th>
-                  <th>Show Details</th>
                 </tr>
                 </thead>
                 <tbody>
-                 @foreach($records as $record)
                   <tr>
-                      <td>{{$loop->iteration}}</td>
                       <td>{{$record->id}}</td>
                       <td>{{$record->name}}</td>
+                      <td>{{$record->phone}}</td>
+                      <td>{{$record->subject}}</td>
+                      <td>{{$record->body}}</td>
                       <td>{{$record->created_at}}</td>
                       <td>{{$record->updated_at}}</td>
                       <td class="text-center">
-                        <a href="{{url(route('category.edit', $record->id))}}" class="btn btn-info btn-xs"><i class="fas fa-edit"></i></a>
-                      </td>
-                      <td class="text-center">
                         {!! Form::open([
-                              'action' => ['CategoryController@destroy', $record->id],
+                              'action' => ['ContactMessageController@destroy', $record->id],
                               'method' => 'delete'
                           ]) !!}
                         <button type="submit" class="btn btn-danger btn-xs">
                           <i class="fas fa-trash"></i>
                         </button> 
                         {!! Form::close() !!}
+
                       </td>
-                      <td>
-                         <a href="{{url(route('category.show', $record->id))}}" class="btn btn-warning btn-xs"><i class="fas fa-list"></i></a>
-                      </td>
-                  </tr>       
-                  @endforeach
-              </table>
-            @else
-              <div class="alert alert-danger" role="alert">
-                  No categories to display.
-              </div>
-            @endif 
+                  </tr>  
+              </table> 
+
+              @endif
+
             </div>
             <!-- /.card-body -->
           </div>
+
+
+          <div class="text-right back-btn">
+            <a href="{{ url()->previous() }}" class="btn btn-primary"><i class="fas fa-arrow-left"></i> &nbsp;Back &nbsp;</a>
+          </div>  
 
     </section>
     <!-- /.content -->
